+++ packages/calcite-components/src/components/chip-group/chip-group.tsx
@@ -12,8 +12,6 @@ import {
 } from "@stencil/core";
 import { focusElementInGroup, slotChangeGetAssignedElements, toAriaBoolean } from "../../utils/dom";
 import {
-  connectInteractive,
-  disconnectInteractive,
   InteractiveComponent,
   InteractiveContainer,
   updateHostInteraction,
@@ -103,12 +101,7 @@ export class ChipGroup implements InteractiveComponent {
   //
   //--------------------------------------------------------------------------
 
-  connectedCallback(): void {
-    connectInteractive(this);
-  }
-
   componentDidRender(): void {
-    disconnectInteractive(this);
     updateHostInteraction(this);
   }
 
@@ -152,11 +145,11 @@ export class ChipGroup implements InteractiveComponent {
     const item = event.target as HTMLCalciteChipElement;
     if (this.items?.includes(item)) {
       if (this.items?.indexOf(item) > 0) {
-        focusElementInGroup(this.items, item as HTMLCalciteChipElement, "previous");
+        focusElementInGroup(this.items, item, "previous");
       } else if (this.items?.indexOf(item) === 0) {
-        focusElementInGroup(this.items, item as HTMLCalciteChipElement, "next");
+        focusElementInGroup(this.items, item, "next");
       } else {
-        focusElementInGroup(this.items, item as HTMLCalciteChipElement, "first");
+        focusElementInGroup(this.items, item, "first");
       }
     }
     this.items = this.items?.filter((el) => el !== item);
@@ -216,11 +209,11 @@ export class ChipGroup implements InteractiveComponent {
   private updateItems = (event?: Event): void => {
     const itemsFromSlot = this.slotRefEl
       ?.assignedElements({ flatten: true })
-      .filter((el) => el?.matches("calcite-chip")) as HTMLCalciteChipElement[];
+      .filter((el): el is HTMLCalciteChipElement => el?.matches("calcite-chip"));
 
     this.items = !event
       ? itemsFromSlot
-      : (slotChangeGetAssignedElements(event) as HTMLCalciteChipElement[]);
+      : slotChangeGetAssignedElements<HTMLCalciteChipElement>(event);
 
     if (this.items?.length < 1) {
       return;