+++ packages/calcite-components/src/components/input-date-picker/input-date-picker.tsx
@@ -44,8 +44,6 @@ import {
   submitForm,
 } from "../../utils/form";
 import {
-  connectInteractive,
-  disconnectInteractive,
   InteractiveComponent,
   InteractiveContainer,
   updateHostInteraction,
@@ -202,7 +200,7 @@ export class InputDatePicker
   }
 
   /**
-   * Defines the available placements that can be used when a flip occurs.
+   * Specifies the component's fallback `calcite-date-picker` `placement` when it's initial or specified `placement` has insufficient space available.
    */
   @Prop() flipPlacements: FlipPlacement[];
 
@@ -388,7 +386,7 @@ export class InputDatePicker
       return;
     }
 
-    const date = dateFromLocalizedString(value, this.localeData) as Date;
+    const date = dateFromLocalizedString(value, this.localeData);
 
     if (inRange(date, this.min, this.max)) {
       this.datePickerActiveDate = date;
@@ -466,7 +464,6 @@ export class InputDatePicker
   // --------------------------------------------------------------------------
 
   connectedCallback(): void {
-    connectInteractive(this);
     connectLocalized(this);
     this.handleDateTimeFormatChange();
 
@@ -493,9 +490,9 @@ export class InputDatePicker
         this.value = "";
       }
     } else if (this.valueAsDate) {
-      if (this.range) {
-        this.setRangeValue(this.valueAsDate as Date[]);
-      } else if (!Array.isArray(this.valueAsDate)) {
+      if (this.range && Array.isArray(this.valueAsDate)) {
+        this.value = [dateToISO(this.valueAsDate[0]), dateToISO(this.valueAsDate[1])];
+      } else if (!this.range && !Array.isArray(this.valueAsDate)) {
         this.value = dateToISO(this.valueAsDate);
       }
     }
@@ -534,7 +531,6 @@ export class InputDatePicker
 
   disconnectedCallback(): void {
     deactivateFocusTrap(this);
-    disconnectInteractive(this);
     disconnectLabel(this);
     disconnectForm(this);
     disconnectFloatingUI(this, this.referenceEl, this.floatingEl);
@@ -904,7 +900,7 @@ export class InputDatePicker
     const { focusedInput, value } = this;
     const focusedInputName = `${focusedInput}Input`;
     const focusedInputValue = this[focusedInputName].value;
-    const date = dateFromLocalizedString(focusedInputValue, this.localeData) as Date;
+    const date = dateFromLocalizedString(focusedInputValue, this.localeData);
     const dateAsISO = dateToISO(date);
     const valueIsArray = Array.isArray(value);
     if (this.range) {
@@ -1142,7 +1138,7 @@ export class InputDatePicker
         this.setInputValue(oldValue[1], "end");
       } else {
         this.value = oldValue;
-        this.setInputValue(oldValue as string);
+        this.setInputValue(oldValue);
       }
     }
   };