+++ packages/calcite-components/src/components/input-time-picker/input-time-picker.tsx
@@ -27,10 +27,7 @@ import {
   MutableValidityState,
   submitForm,
 } from "../../utils/form";
-import { guid } from "../../utils/guid";
 import {
-  connectInteractive,
-  disconnectInteractive,
   InteractiveComponent,
   InteractiveContainer,
   updateHostInteraction,
@@ -366,6 +363,8 @@ export class InputTimePicker
 
   @Element() el: HTMLCalciteInputTimePickerElement;
 
+  @State() calciteInputEl: HTMLCalciteInputElement;
+
   defaultValue: InputTimePicker["value"];
 
   formEl: HTMLFormElement;
@@ -374,22 +373,16 @@ export class InputTimePicker
 
   popoverEl: HTMLCalcitePopoverElement;
 
-  private calciteInputEl: HTMLCalciteInputElement;
-
   private calciteTimePickerEl: HTMLCalciteTimePickerElement;
 
   private focusOnOpen = false;
 
   focusTrap: FocusTrap;
 
-  private dialogId = `time-picker-dialog--${guid()}`;
-
   private localeConfig: ILocale;
 
   /** whether the value of the input was changed as a result of user typing or not */
   private userChangedValue = false;
-
-  private referenceElementId = `input-time-picker-${guid()}`;
 
   //--------------------------------------------------------------------------
   //
@@ -857,6 +850,7 @@ export class InputTimePicker
 
   private setCalcitePopoverEl = (el: HTMLCalcitePopoverElement): void => {
     this.popoverEl = el;
+    this.openHandler();
   };
 
   private setInputEl = (el: HTMLCalciteInputElement): void => {
@@ -952,7 +946,6 @@ export class InputTimePicker
   //--------------------------------------------------------------------------
 
   connectedCallback() {
-    connectInteractive(this);
     connectLocalized(this);
 
     if (isValidTime(this.value)) {
@@ -984,12 +977,9 @@ export class InputTimePicker
         }),
       );
     }
-
-    this.openHandler();
   }
 
   disconnectedCallback() {
-    disconnectInteractive(this);
     disconnectLabel(this);
     disconnectForm(this);
     disconnectLocalized(this);
@@ -1008,7 +998,7 @@ export class InputTimePicker
   // --------------------------------------------------------------------------
 
   render(): VNode {
-    const { disabled, messages, readOnly, dialogId } = this;
+    const { disabled, messages, readOnly } = this;
     return (
       <Host onBlur={this.hostBlurHandler} onKeyDown={this.keyDownHandler}>
         <InteractiveContainer disabled={this.disabled}>
@@ -1020,7 +1010,6 @@ export class InputTimePicker
               aria-invalid={toAriaBoolean(this.status === "invalid")}
               disabled={disabled}
               icon="clock"
-              id={this.referenceElementId}
               label={getLabelText(this)}
               lang={this.effectiveLocale}
               onCalciteInputTextInput={this.calciteInternalInputInputHandler}
@@ -1030,12 +1019,13 @@ export class InputTimePicker
               role="combobox"
               scale={this.scale}
               status={this.status}
-            />
-            {!this.readOnly && this.renderToggleIcon(this.open)}
+            >
+              {!this.readOnly && this.renderToggleIcon(this.open)}
+            </calcite-input-text>
           </div>
           <calcite-popover
+            autoClose={true}
             focusTrapDisabled={true}
-            id={dialogId}
             label={messages.chooseTime}
             lang={this.effectiveLocale}
             onCalcitePopoverBeforeClose={this.popoverBeforeCloseHandler}
@@ -1045,7 +1035,7 @@ export class InputTimePicker
             overlayPositioning={this.overlayPositioning}
             placement={this.placement}
             ref={this.setCalcitePopoverEl}
-            referenceElement={this.referenceElementId}
+            referenceElement={this.calciteInputEl}
             triggerDisabled={true}
           >
             <calcite-time-picker
@@ -1077,7 +1067,7 @@ export class InputTimePicker
 
   renderToggleIcon(open: boolean): VNode {
     return (
-      <span class={CSS.toggleIcon}>
+      <span class={CSS.toggleIcon} slot="action">
         <calcite-icon
           icon={open ? "chevron-up" : "chevron-down"}
           scale={getIconScale(this.scale)}