+++ packages/calcite-components/src/components/select/select.tsx
@@ -21,8 +21,6 @@ import {
   MutableValidityState,
 } from "../../utils/form";
 import {
-  connectInteractive,
-  disconnectInteractive,
   InteractiveComponent,
   InteractiveContainer,
   updateHostInteraction,
@@ -193,14 +191,12 @@ export class Select
       childList: true,
     });
 
-    connectInteractive(this);
     connectLabel(this);
     connectForm(this);
   }
 
   disconnectedCallback(): void {
     this.mutationObserver?.disconnect();
-    disconnectInteractive(this);
     disconnectLabel(this);
     disconnectForm(this);
   }
@@ -345,7 +341,7 @@ export class Select
       if (isOption(optionOrGroup) && nativeOptionOrGroup === nativeOption) {
         optionOrGroup.selected = true;
         futureSelected = optionOrGroup;
-        this.deselectAllExcept(optionOrGroup as HTMLCalciteOptionElement);
+        this.deselectAllExcept(optionOrGroup);
       }
     });