+++ packages/calcite-ui-icons/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.32.0-next.2](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@3.32.0-next.1...@esri/calcite-ui-icons@3.32.0-next.2) (2024-09-11)
+
+### Features
+
+- add justify ([#10272](https://github.com/Esri/calcite-design-system/issues/10272)) ([3d6c013](https://github.com/Esri/calcite-design-system/commit/3d6c013d193403cf7f5c669150f62266c919c7dc))
+
+## [3.32.0-next.1](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@3.32.0-next.0...@esri/calcite-ui-icons@3.32.0-next.1) (2024-08-30)
+
+### Features
+
+- added connection end left, connection end right, connection middle, connection to connection ([#10198](https://github.com/Esri/calcite-design-system/issues/10198)) ([3d0f703](https://github.com/Esri/calcite-design-system/commit/3d0f703e552db63d4b44b67e7f2590336d227f7c))
+
+## [3.32.0-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@3.31.1-next.0...@esri/calcite-ui-icons@3.32.0-next.0) (2024-08-30)
+
+### Features
+
+- add trace path and trace path complete ([#10196](https://github.com/Esri/calcite-design-system/issues/10196)) ([c37ced5](https://github.com/Esri/calcite-design-system/commit/c37ced57fc08884639ed3cacafa2023104a33696))
+
+## [3.31.1-next.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@3.31.0-next.4...@esri/calcite-ui-icons@3.31.1-next.0) (2024-08-28)
+
+**Note:** Version bump only for package @esri/calcite-ui-icons
+
 ## [3.31.0](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-ui-icons@3.30.0...@esri/calcite-ui-icons@3.31.0) (2024-08-28)
 
 ### Features